# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for local RAG indexing and document management",
        "title": "Local RAG API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/discussions": {
            "get": {
                "description": "Retrieves a list of all discussions with their summaries",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "List all discussions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-array_discussion_DiscussionSummaryResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new discussion with the specified title",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Create a new discussion",
                "parameters": [
                    {
                        "description": "Discussion creation parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateDiscussionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Response-discussion_CreateDiscussionResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/discussions/{id}": {
            "get": {
                "description": "Retrieves a complete discussion including all messages",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Get a discussion by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discussion ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-discussion_DiscussionResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/discussions/{id}/history": {
            "get": {
                "description": "Retrieves all messages in a discussion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Get discussion message history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discussion ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.Response-discussion_GetDiscussionHistoryResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/discussions/{id}/question": {
            "post": {
                "description": "Adds a user question to a discussion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Ask a question in a discussion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discussion ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Question parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AskQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Response-discussion_AskQuestionResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/discussions/{id}/question/stream": {
            "post": {
                "description": "Adds a user question to a discussion and streams the AI response using Server-Sent Events",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Ask a question with streaming response",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Discussion ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Question parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AskQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Streaming response (SSE stream of tokens)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
        "/questions": {
            "post": {
                "description": "Asks a question, creating a new discussion if discussion_id is not provided. Uses mistral:latest to auto-generate discussion titles.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Ask a question (with auto-discussion creation)",
                "parameters": [
                    {
                        "description": "Question parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.QuickQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.Response-discussion_AskQuestionQuickResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
        "/questions/stream": {
            "post": {
                "description": "Asks a question with streaming response, creating a new discussion if discussion_id is not provided",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Ask a question with streaming (with auto-discussion creation)",
                "parameters": [
                    {
                        "description": "Question parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.QuickQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Streaming response (SSE stream of tokens)",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response-any"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.AskQuestionRequest": {
            "type": "object",
            "required": [
                "question"
            ],
            "properties": {
                "question": {
                    "type": "string",
                    "example": "What are vector embeddings?"
                }
            }
        },
        "api.CreateDiscussionRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "example": "How to use embeddings?"
                }
            }
        },
...
        "api.QuickQuestionRequest": {
            "type": "object",
            "required": [
                "question"
            ],
            "properties": {
                "discussion_id": {
                    "type": "string",
                    "example": ""
                },
                "question": {
                    "type": "string",
                    "example": "What are vector embeddings?"
                }
            }
        },
        "api.Response-any": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message (omitted on success)",
                    "type": "string"
                },
                "result": {
                    "description": "The response data (omitted on error)"
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "api.Response-array_discussion_DiscussionSummaryResult": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message (omitted on success)",
                    "type": "string"
                },
                "result": {
                    "description": "The response data (omitted on error)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/discussion.DiscussionSummaryResult"
                    }
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "api.Response-discussion_AskQuestionQuickResult": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message (omitted on success)",
                    "type": "string"
                },
                "result": {
                    "description": "The response data (omitted on error)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/discussion.AskQuestionQuickResult"
                        }
                    ]
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "api.Response-discussion_AskQuestionResult": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message (omitted on success)",
                    "type": "string"
                },
                "result": {
                    "description": "The response data (omitted on error)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/discussion.AskQuestionResult"
                        }
                    ]
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "api.Response-discussion_CreateDiscussionResult": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message (omitted on success)",
                    "type": "string"
                },
                "result": {
                    "description": "The response data (omitted on error)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/discussion.CreateDiscussionResult"
                        }
                    ]
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "api.Response-discussion_DiscussionResult": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message (omitted on success)",
                    "type": "string"
                },
                "result": {
                    "description": "The response data (omitted on error)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/discussion.DiscussionResult"
                        }
                    ]
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "api.Response-discussion_GetDiscussionHistoryResult": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message (omitted on success)",
                    "type": "string"
                },
                "result": {
                    "description": "The response data (omitted on error)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/discussion.GetDiscussionHistoryResult"
                        }
                    ]
                },
                "status": {
                    "description": "\"success\" or \"error\"",
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "discussion.CreateDiscussionResult": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "discussion.DiscussionResult": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/discussion.MessageResult"
                    }
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "discussion.DiscussionSummaryResult": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "message_count": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "discussion.GetDiscussionHistoryResult": {
            "type": "object",
            "properties": {
                "discussion_id": {
                    "type": "string"
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/discussion.MessageResult"
                    }
                }
            }
        },
        "discussion.MessageResult": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
    }
}